requirements.txt
python-dotenv==1.0.0
PyYAML==6.0
requests==2.27.1
